requirements.txt
llangchain>=1.2.7
mistralai>=0.0.12
langchain-mistralai>=0.0.3
openai>=1.0.0
langchain-openai>=0.1.0
google-genai>=0.1.0
langchain-google-genai>=0.1.0

faiss-cpu>=1.13.2
pandas>=3.0.0
dotenv>=0.9.9
fastapi>=0.128.0
uvicorn>=0.40.0